Upgrade to stats-ag v0.1.2

+++ recipes/default.rb
@@ -58,21 +58,24 @@ execute 'build & install stats-ag binary' do
     export GOPATH=#{node['go']['gopath']}
     export GOBIN=#{node['go']['gobin']}
     #{node['go']['install_dir']}/go/bin/go get github.com/shirou/gopsutil
-    #{node['go']['install_dir']}/go/bin/go build -o bin/stats-ag-#{node['stats_ag']['git_tag']} -ldflags "-X main.BUILD_DATE `date +%Y-%m-%d` -X main.VERSION #{node['stats_ag']['git_tag']} -X main.COMMIT_SHA `git rev-parse --verify HEAD`"
-    mv bin/stats-ag-#{node['stats_ag']['git_tag']} #{node['stats_ag']['base_dir']}/stats-ag-#{node['stats_ag']['git_tag']}
-    unlink #{node['stats_ag']['base_dir']}/stats-ag  
-    unlink /usr/bin/stats-ag
-    ln -s #{node['stats_ag']['base_dir']}/stats-ag-#{node['stats_ag']['git_tag']} #{node['stats_ag']['base_dir']}/stats-ag
-    ln -s #{node['stats_ag']['base_dir']}/stats-ag-#{node['stats_ag']['git_tag']} /usr/bin/stats-ag
+    #{node['go']['install_dir']}/go/bin/go build -o bin/stats-ag -ldflags "-X main.BUILD_DATE `date +%Y-%m-%d` -X main.VERSION #{node['stats_ag']['git_tag']} -X main.COMMIT_SHA `git rev-parse --verify HEAD`"
+    if [ -e /usr/bin/stats-ag ]; then
+      unlink /usr/bin/stats-ag
+    fi
+    if [ -e /usr/bin/stats-ag ]; then
+      rm #{node['stats_ag']['base_dir']}/stats-ag && cp bin/stats-ag #{node['stats_ag']['base_dir']}/stats-ag
+    else
+      cp bin/stats-ag #{node['stats_ag']['base_dir']}/stats-ag
+    fi
+    ln -s #{node['stats_ag']['base_dir']}/stats-ag /usr/bin/stats-ag
   EOF
   action :run
-  not_if { File.exist?("#{node['stats_ag']['base_dir']}/stats-ag-#{node['stats_ag']['git_tag']}") }
 end
 
 cron 'run stats-ag every minute' do
   hour '*'
   minute '*'
-  command "#{node['stats_ag']['base_dir']}/stats-ag -e 1 -m #{node['stats_ag']['metrics_dir']} -s #{node['stats_ag']['scripts_dir']} -p #{node['stats_ag']['date_prefix_format']} > #{node['stats_ag']['log_file']} 2>&1"
+  command "#{node['stats_ag']['base_dir']}/stats-ag -m #{node['stats_ag']['metrics_dir']} -s #{node['stats_ag']['scripts_dir']} -p #{node['stats_ag']['date_prefix_format']} > #{node['stats_ag']['log_file']} 2>&1"
 end
 
 metrics_list.each do |type|
